refact(desk): moved Desk into its own file
